move hoog useGetParameters and create new rowAction
