+++ Client/src/components/Parameters/ParameterDetails.tsx
@@ -1,0 +1,11 @@
+
+
+const ParameterDetails = () => {
+    return (
+        <div>
+            
+        </div>
+    );
+};
+
+export default ParameterDetails;
