+++ Client/src/pages/ParametersPage.tsx
@@ -1,8 +1,9 @@
 import TableLayout from "./layouts/TableLayout";
 import ComponentOrLoader from "@components/Shared/ComponentOrLoader.tsx";
 import Loader from "@components/Shared/Loader.tsx";
-import useGetParameters from "@hooks/useGetParameters.ts";
+import useGetParameters from "@hooks/parameters/useGetParameters.ts";
 import Parameters from "@components/Parameters/Parameters.tsx";
+import ParameterDetails from "@components/Parameters/ParameterDetails.tsx";
 
 const ParametersPage = () => {
 
@@ -18,16 +19,16 @@ const ParametersPage = () => {
           parameters={dataParameters?.data?.parameters || []}
         />
       </ComponentOrLoader>
-      {/*<ComponentOrLoader*/}
-      {/*  isLoading={dataQuerySampleDetails.isLoading}*/}
-      {/*  loader={<Loader />}*/}
-      {/*>*/}
-      {/*  <SampleDetails*/}
-      {/*    sample={dataQuerySampleDetails.data}*/}
-      {/*    open={detailsOpen}*/}
-      {/*    openChange={setDetailsOpen}*/}
-      {/*  />*/}
-      {/*</ComponentOrLoader>*/}
+      <ComponentOrLoader
+        isLoading={dataParameters.isLoading}
+        loader={<Loader />}
+      >
+        <ParameterDetails
+          // sample={dataParameters.data}
+          // open={detailsOpen}
+          // openChange={setDetailsOpen}
+        />
+      </ComponentOrLoader>
     </TableLayout>
   );
 };
